backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-cloud-aiplatform==1.42.1
pydantic==2.5.3
